refactor: consolidate defaultSendActionRuntime import to network module
