+++ packages/insomnia/src/common/send-request.ts
@@ -14,7 +14,7 @@ import {
   tryToExecutePreRequestScript,
   tryToInterpolateRequest,
 } from '../network/network';
-import { defaultSendActionRuntime } from '../ui/routes/$organizationId.project.$projectId.workspace.$workspaceId.debug.request.$requestId';
+import { defaultSendActionRuntime } from '../network/network';
 import { database } from './database';
 
 // The network layer uses settings from the settings model
